Medical info on each birth in the US in 2014

When using Web-R, load this dataset with load("https://www.mosaic-web.org/go/datasets/Natality_2014.rda")

Usage

Natality_2014

Format

A data frame with a random sample of size 100000 from the complete CDC set of 3,998,175 cases, each of which is a birth in the US in 2014.

mager Mother's age at date of birth

fagecomb Father's age at date of birth

ubfacil Type of facility where the birth took place

urf_diab Mother had diabetes before pregnancy

urf_chyper Mother had chronic hypertension before pregnancy

urf_phyper Mother developed hypertension during pregnancy

urf_eclam Mother developed eclampsia

uop_induc Was labor induced?

uld_breech Did baby present as breech? See me_pres.

ilive Whether the baby was alive at the time of the report.

ab_aven1 Baby put on mechanical ventilator immediately

ab_aven6 Baby still on ventilator after 6 hours

ab_nicu Baby transferred to intensive care unit.

ab_surf Surfactant administered to newborn

ab_anti Antibiotics administered to newborn

ab_seiz Newborn had seizure

dbwt Baby's weight (gm)

combgest Length of gestation

sex Baby's sex

dplural Plurality of birth: 2 = twins, 3 = triplets, ...

apgar5 APGAR score at 5 minutes

apgar10 APGAR score at 10 minutes

mtran Mother transferred. No other description given.

pay Source of payment for delivery

mm_mtr Mother received blood transfusion

mm_plac Perineal laceraction

mm_rupt Ruptured uterus

mm_uhyst Unplanned hysterectomy

mm_aicu Mother transferred to intensive care unit

me_pres Presentation of baby on delivery. See uld_breech. It's not clear why "other" corresponds to breech.

me_rout Method of delivery

ld_indl Labor induced

pwgt_r Mother's weight before pregnancy

dwgt_r Mother's weight at delivery

m_ht_in Mother's height

cig_0 Number of cigarettes smoked daily before pregnancy

cig_1 Number of cigarettes smoked daily during first trimester

cig_2 Number of cigarettes smoked daily during second trimester

cig_3 Number of cigarettes smoked daily during third trimester

wic Enrolled in Women, Infants, and Children (WIC) program for supplemental nutrition.

precare Month started in prenatal care. 15 means never started.

previs Number of visits to prenatal care.

priorlive Number of previous live births.

priordead Number of previous births where baby died.

priorterm Number of terminations of pregnancy before this birth.

Details

The Centers for Disease Control collects data on all births registered in the US (50 states + DC). The full data set, Natality_2014, contains entries for the full set of births keeping only a subset of variables related to potential risk factors and outcomes. Natality_2014 is a random sample (specifically for the Little Apps) of size ten thousand. A larger sample of one-million cases, Natality_2014_1000k, and the complete set of cases in Natality_2014 are also provided by the package, but not through the usual data() interface. To access those large files, see the instructions given in Larger_natality_data_files.

Variable names are the same as in the original CDC file.

Variables have been re-coded to translate markers like 99 into NA and to self-document the levels of the variables.

Note that the data include only registered births. Fetal deaths are not included. For reference, there were about 24,000 fetal deaths (20 weeks and thereafter) in the US in 2013. (See http://www.cdc.gov/nchs/data/nvsr/nvsr64/nvsr64_08.pdf.) This corresponds to about 0.6% of the births listed in these data. The CDC also reports about 16,000 infant deaths before age 28 days and another 7500 deaths before age 1 year.

Larger data files are available by following the instructions in Larger_natality_data_files in the Natality_2014 package available for R: Natality_2014 with 3.99 million cases and Natality_2014_1000k with a random sample of one million cases.

See also