Mix two variables together. The output will have the specified R-squared with var1 and variance one.

Mix two variables together. The output will have the specified R-squared with var1 and variance one.

Usage

mix_with(signal, noise = NULL, R2 = 0.5, var = 1, exact = FALSE)

Arguments

signal

The part of the mixture that will be correlated with the output.

noise

The rest of the mixture. This will be uncorrelated with the output only if you specify it as pure noise.

R2

The target R-squared.

var

The target variance.

exact

if TRUE, make R-squared or the target variance exactly as specified.

Details

The target R-squared and variance will be achieved only if exact=TRUE or the sample size goes to infinity.