Check model type against model specification and data

This can be used to automatically determine a model type or to determine if the specified model type is consistent with the specification/data

Usage

model_family(
  .data,
  .tilde,
  family = c("auto", "lm", "linear", "binomial", "poisson", "svm", "gaussian", "rlm")
)

Arguments

.data

A data frame or equivalent

.tilde

A model specification as a tilde expression

family

Requested model type, if any.