create columns with random numbers

For demonstration purposes, add the specified number of random columns to a model matrix. This is intended to be used in modeling functions, e.g. model_train(), lm(), and so on to explore the extent to which random columns "explain" the response variable.

Usage

random_terms(df = 1, rdist = rnorm, args = list(), n, seed = NULL)

Arguments

df

How many columns to add

rdist

Function to generate each column's numbers (default: rnorm)

args

A list holding the parameters (if any) to be used for the rdist argument

n

OPTIONALLY, dictate the number of rows in the output

seed

Integer seed for the random-number generator

Details

random_terms() will try to guess a suitable value for n based on the calling function.

Examples

 mtcars |> model_train(mpg ~ wt + random_terms(4)) |> conf_interval()
#> == Building pkgdown site =======================================================
#> Reading from: '/Users/kaplan/Packages/LSTbook'
#> Writing to:   '/Users/kaplan/Packages/LSTbook/docs'
#> -- Initialising site -----------------------------------------------------------
#> -- Building home ---------------------------------------------------------------
#> Reading 'LICENSE.md'
#> Writing '404.html'
#> -- Building function reference -------------------------------------------------
#> Reading 'man/AAUP.Rd'
#> Reading 'man/Anthro_F.Rd'
#> Reading 'man/Birdkeepers.Rd'
#> Reading 'man/Births2022.Rd'
#> Reading 'man/Boston_marathon.Rd'
#> Reading 'man/Butterfly.Rd'
#> Reading 'man/CRDS.Rd'
#> Reading 'man/Calif_precip.Rd'
#> Reading 'man/Callback.Rd'
#> Reading 'man/College_grades.Rd'
#> Reading 'man/Econ_outlook_poll.Rd'
#> Reading 'man/FARS.Rd'
#> Reading 'man/Framingham.Rd'
#> Reading 'man/Germany1933vote.Rd'
#> Reading 'man/Gilbert.Rd'
#> Reading 'man/Go_vote.Rd'
#> Reading 'man/Hill_racing.Rd'
#> Reading 'man/MPG.Rd'
#> Reading 'man/McClave_Sincich.Rd'
#> Reading 'man/McCredie_Kurtz.Rd'
#> Reading 'man/Monocacy_river.Rd'
#> Reading 'man/Offspring.Rd'
#> Reading 'man/Orings.Rd'
#> Reading 'man/PIDD.Rd'
#> Reading 'man/Penguins.Rd'
#> Reading 'man/Registrar.Rd'
#> Reading 'man/STAR.Rd'
#> Reading 'man/Shipping_losses.Rd'
#> Reading 'man/Short_examples.Rd'
#> Reading 'man/UCB_applicants.Rd'
#> Reading 'man/US_wildfires.Rd'
#> Reading 'man/Wheat.Rd'
#> Reading 'man/add_plot_labels.Rd'
#> Error in purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,     examples_env = examples_env, run_dont_run = run_dont_run): ℹ In index: 33.
#> ℹ With name: add_plot_labels.Rd.
#> Caused by error in `.f()`:
#> ! Failed to parse Rd in add_plot_labels.Rd
#> ℹ sink stack is full
#> Caused by error in `sink()`:
#> ! sink stack is full
 mtcars |> model_train(mpg ~ wt + random_terms(4)) |> anova_summary()
#> == Building pkgdown site =======================================================
#> Reading from: '/Users/kaplan/Packages/LSTbook'
#> Writing to:   '/Users/kaplan/Packages/LSTbook/docs'
#> -- Initialising site -----------------------------------------------------------
#> -- Building home ---------------------------------------------------------------
#> Reading 'LICENSE.md'
#> Writing '404.html'
#> -- Building function reference -------------------------------------------------
#> Reading 'man/AAUP.Rd'
#> Reading 'man/Anthro_F.Rd'
#> Reading 'man/Birdkeepers.Rd'
#> Reading 'man/Births2022.Rd'
#> Reading 'man/Boston_marathon.Rd'
#> Reading 'man/Butterfly.Rd'
#> Reading 'man/CRDS.Rd'
#> Reading 'man/Calif_precip.Rd'
#> Reading 'man/Callback.Rd'
#> Reading 'man/College_grades.Rd'
#> Reading 'man/Econ_outlook_poll.Rd'
#> Reading 'man/FARS.Rd'
#> Reading 'man/Framingham.Rd'
#> Reading 'man/Germany1933vote.Rd'
#> Reading 'man/Gilbert.Rd'
#> Reading 'man/Go_vote.Rd'
#> Reading 'man/Hill_racing.Rd'
#> Reading 'man/MPG.Rd'
#> Reading 'man/McClave_Sincich.Rd'
#> Reading 'man/McCredie_Kurtz.Rd'
#> Reading 'man/Monocacy_river.Rd'
#> Reading 'man/Offspring.Rd'
#> Reading 'man/Orings.Rd'
#> Reading 'man/PIDD.Rd'
#> Reading 'man/Penguins.Rd'
#> Reading 'man/Registrar.Rd'
#> Reading 'man/STAR.Rd'
#> Reading 'man/Shipping_losses.Rd'
#> Reading 'man/Short_examples.Rd'
#> Reading 'man/UCB_applicants.Rd'
#> Reading 'man/US_wildfires.Rd'
#> Reading 'man/Wheat.Rd'
#> Reading 'man/add_plot_labels.Rd'
#> Error in purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,     examples_env = examples_env, run_dont_run = run_dont_run): ℹ In index: 33.
#> ℹ With name: add_plot_labels.Rd.
#> Caused by error in `.f()`:
#> ! Failed to parse Rd in add_plot_labels.Rd
#> ℹ sink stack is full
#> Caused by error in `sink()`:
#> ! sink stack is full
 head(mtcars) |> select(wt, mpg) |> mutate(r = random_terms(3))
#> == Building pkgdown site =======================================================
#> Reading from: '/Users/kaplan/Packages/LSTbook'
#> Writing to:   '/Users/kaplan/Packages/LSTbook/docs'
#> -- Initialising site -----------------------------------------------------------
#> -- Building home ---------------------------------------------------------------
#> Reading 'LICENSE.md'
#> Writing '404.html'
#> -- Building function reference -------------------------------------------------
#> Reading 'man/AAUP.Rd'
#> Reading 'man/Anthro_F.Rd'
#> Reading 'man/Birdkeepers.Rd'
#> Reading 'man/Births2022.Rd'
#> Reading 'man/Boston_marathon.Rd'
#> Reading 'man/Butterfly.Rd'
#> Reading 'man/CRDS.Rd'
#> Reading 'man/Calif_precip.Rd'
#> Reading 'man/Callback.Rd'
#> Reading 'man/College_grades.Rd'
#> Reading 'man/Econ_outlook_poll.Rd'
#> Reading 'man/FARS.Rd'
#> Reading 'man/Framingham.Rd'
#> Reading 'man/Germany1933vote.Rd'
#> Reading 'man/Gilbert.Rd'
#> Reading 'man/Go_vote.Rd'
#> Reading 'man/Hill_racing.Rd'
#> Reading 'man/MPG.Rd'
#> Reading 'man/McClave_Sincich.Rd'
#> Reading 'man/McCredie_Kurtz.Rd'
#> Reading 'man/Monocacy_river.Rd'
#> Reading 'man/Offspring.Rd'
#> Reading 'man/Orings.Rd'
#> Reading 'man/PIDD.Rd'
#> Reading 'man/Penguins.Rd'
#> Reading 'man/Registrar.Rd'
#> Reading 'man/STAR.Rd'
#> Reading 'man/Shipping_losses.Rd'
#> Reading 'man/Short_examples.Rd'
#> Reading 'man/UCB_applicants.Rd'
#> Reading 'man/US_wildfires.Rd'
#> Reading 'man/Wheat.Rd'
#> Reading 'man/add_plot_labels.Rd'
#> Error in mutate(select(head(mtcars), wt, mpg), r = random_terms(3)): ℹ In argument: `r = random_terms(3)`.
#> Caused by error in `purrr::map()`:
#> ℹ In index: 33.
#> ℹ With name: add_plot_labels.Rd.
#> Caused by error in `.f()`:
#> ! Failed to parse Rd in add_plot_labels.Rd
#> ℹ sink stack is full
#> Caused by error in `sink()`:
#> ! sink stack is full