Run the left side of the pipeline multiple times.

Write a pipeline to perform some calculation whose result can be coerced into one line of a data frame. Add trials(times=3) to the end of the pipeline in order to repeat the calculation multiple times. Typically, each trial involves some random component, but another (or an additional) capability is to parameterize the pipeline expression by including some unbound variable in it, e.g. lambda. Then call trials(lambda=c(10,20)) to repeat the calculation for each of the elements of the named parameter.

Usage

trials(.ex, times = 1, ...)

Arguments

.ex

(Not user-facing.) The left side of the pipeline.

times

The number of times to run the trial.

...

Values for any unbound parameter in the left side of the pipeline. If a vector of length > 1, the trials will be run separately for each element of the vector.

Value

a dataframe with one row for each trial. (But see the ... argument.)

Details

This is intended as a pipeline friendly replacement for mosaic::do().

Examples

mean(rnorm(10)) |> trials(times=3)
#> New names:
#> • `` -> `...2`
#> New names:
#> • `` -> `...2`
#> New names:
#> • `` -> `...2`
#>   .trial       ...2
#> 1      1  0.1355040
#> 2      2  0.7796057
#> 3      3 -0.1295242
mean(rnorm(lambda)) |> trials(lambda=c(1, 100, 10000))
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#>   .trial lambda         ...3
#> 1      1      1  1.078517525
#> 2      1    100  0.027760336
#> 3      1  10000 -0.005564744
mean(rnorm(lambda)) |> trials(times=5, lambda=c(1, 100, 10000))
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#> New names:
#> • `` -> `...3`
#>    .trial lambda         ...3
#> 1       1      1 -0.546449966
#> 2       2      1  0.604958074
#> 3       3      1  0.500036123
#> 4       4      1  0.388217185
#> 5       5      1  0.852351189
#> 6       1    100 -0.084561838
#> 7       2    100 -0.068460779
#> 8       3    100 -0.154239487
#> 9       4    100 -0.060918370
#> 10      5    100 -0.186271712
#> 11      1  10000  0.007646683
#> 12      2  10000 -0.020246237
#> 13      3  10000  0.008982420
#> 14      4  10000  0.002411333
#> 15      5  10000  0.015156344
take_sample(mtcars, n=lambda, replace=TRUE) |> select(mpg, hp) |>
  model_train(mpg ~ resample(hp)) |>
  regression_summary() |> trials(times=3, lambda=c(10, 20, 40)) |>
  filter(term == "resample(hp)")
#> Warning: The `tidy()` method for objects of class `model_object` is not maintained by the broom team, and is only supported through the `lm` tidier method. Please be cautious in interpreting and reporting broom output.
#> 
#> This warning is displayed once per session.
#>   .trial lambda         term     estimate  std.error  statistic     p.value
#> 1      1     10 resample(hp)  0.098746917 0.02432590  4.0593329 0.003636804
#> 2      2     10 resample(hp) -0.045460777 0.03585710 -1.2678318 0.240511254
#> 3      3     10 resample(hp) -0.081086576 0.04491934 -1.8051596 0.108692294
#> 4      1     20 resample(hp) -0.015708407 0.04168843 -0.3768050 0.710723785
#> 5      2     20 resample(hp) -0.010002710 0.01836100 -0.5447803 0.592590854
#> 6      3     20 resample(hp)  0.012678416 0.01519494  0.8343843 0.415001324
#> 7      1     40 resample(hp)  0.007472809 0.01342789  0.5565139 0.581122395
#> 8      2     40 resample(hp) -0.012123982 0.01830818 -0.6622168 0.511828733
#> 9      3     40 resample(hp) -0.029782613 0.01142542 -2.6066978 0.012988264